refactor: migrate kill moves to typescript
